+++ tests/test_zk.rs
@@ -6,11 +6,11 @@ use zookeeper_async::{Acl, CreateMode, WatchedEvent, ZooKeeper};
 use test::ZkCluster;
 
 use env_logger;
-use log::*;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
+use tracing::*;
 
 #[tokio::test]
 async fn zk_test() {